refactor(domain): split filter_config.py into filtering submodule
Break down 354-line filter_config.py into smaller, focused modules:
- column_filter.py (28 lines): GoldColumnFilter
- range_filter.py (36 lines): GoldRangeFilter
- list_filters.py (55 lines): GoldListLengthFilter, GoldListContainsFilter
- gold_config.py (185 lines): GoldFilterConfig with all validation logic
- input_config.py (51 lines): InputFilterConfig
- load_result.py (44 lines): FilterLoadResult
- __init__.py (28 lines): Public API facade

All files now under 200 lines. Backwards compatible via re-exports
in filter_config.py and domain/__init__.py.
