+++ src/bioetl/domain/filtering/input_config.py
@@ -1,0 +1,51 @@
+"""Input filter configuration.
+
+Provides configuration for input-based filtering of pipeline records.
+"""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class InputFilterConfig:
+    """Configuration for input-based filtering.
+
+    When enabled, the pipeline will only fetch records matching
+    the IDs provided in the source file.
+
+    Attributes:
+        enabled: Whether filtering is active.
+        source_path: Path to the filter source (e.g., CSV file).
+        column_name: Name of the column containing filter IDs.
+        filter_field: API field to filter by (e.g., molecule_chembl_id).
+        batch_size: Number of IDs per API request (ChEMBL limit ~100).
+    """
+
+    enabled: bool = False
+    source_path: str | None = None
+    column_name: str | None = None
+    filter_field: str | None = None
+    batch_size: int = 100
+
+    def __post_init__(self) -> None:
+        """Validate configuration consistency."""
+        self._validate_enabled_fields()
+        self._validate_batch_size()
+
+    def _validate_enabled_fields(self) -> None:
+        """Validate fields required when filtering is enabled."""
+        if not self.enabled:
+            return
+        if not self.source_path:
+            raise ValueError("source_path is required when filter is enabled")
+        if not self.column_name:
+            raise ValueError("column_name is required when filter is enabled")
+        if not self.filter_field:
+            raise ValueError("filter_field is required when filter is enabled")
+
+    def _validate_batch_size(self) -> None:
+        """Validate the batch_size is within a reasonable range."""
+        if not (1 <= self.batch_size <= 1000):
+            raise ValueError("batch_size must be between 1 and 1000")
